[Fix] 리뷰서비스 수정
특정 유저의 리뷰를 불러올 때, 해당 리뷰가 참조하는 코스의 순차번호와 코스명을 같이 불러오도록 수정

+++ src/main/java/com/mysite/travelo/gil/review/ReviewRepository.java
@@ -10,7 +10,7 @@ public interface ReviewRepository extends JpaRepository<Review, Integer> {
 	Page<Review> findByCourseCourseSeq(Pageable pageable, Integer courseSeq);
 	
 //	특정 코스의 댓글 개수 조회
-    long countByCourseCourseSeq(Integer courseSeq);
+    int countByCourseCourseSeq(Integer courseSeq);
 	
 //	특정 유저의 댓글 전체보기(정렬 디폴트값: 최신순 / 옵션값: 오래된순)
 	Page<Review> findByUserUserSeq(Pageable pageable, Integer userSeq);
